% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_lapply_functions.R
\name{calculate_parallel_instances}
\alias{calculate_parallel_instances}
\title{Calculate Parallel Instances}
\usage{
calculate_parallel_instances(buffer_mb = 500)
}
\arguments{
\item{buffer_mb}{Memory buffer in MB to subtract from available memory (default is 500 MB).}
}
\value{
The maximum number of parallel instances that can be run.
}
\description{
This function calculates the number of parallel instances that can be run based on the current memory usage.
}
\examples{
calculate_parallel_instances()
}
