% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_lapply_functions.R
\name{smart.lapply}
\alias{smart.lapply}
\title{Smart Apply Function}
\usage{
smart.lapply(
  vector,
  code_function,
  cores = 1,
  seeds_vec = NULL,
  mem_check = TRUE,
  generate_seeds = FALSE
)
}
\arguments{
\item{vector}{The vector to process.}

\item{code_function}{The function to apply to each element of the vector.}

\item{cores}{The number of cores to use.}

\item{seeds_vec}{Optional vector of seeds for reproducibility.}

\item{mem_check}{If \code{TRUE}, check memory pool to limit instances to avoid overflow.}

\item{generate_seeds}{If \code{TRUE}, pre-generate a random set of seeds of \code{length(vector)} for pseudo-random number generation.}
}
\description{
This function applies a function to each element of a vector in parallel, with memory checks.
}
\examples{
\dontrun{
# behaves like standard lapply
smart.lapply(1:100, function(x) x^2)

# parallel version
smart.lapply(1:4, function(x) x^2, cores = 4)

# parallel version, with default to the number of available cores - 2
smart.lapply(1:4, function(x) x^2, cores = 400)

# random number generation without seeds
smart.lapply(1:4, function(x) rnorm(1, 20, 1), cores = 4)

# random number generation with seeds automatically created
smart.lapply(1:4, function(x) rnorm(1, 20, 1), cores = 4, generate_seeds = TRUE)
smart.lapply(1:4, function(x) rnorm(1, 20, 1), cores = 4,
seeds_vec = c(111111111, 111111111, 222222222, 222222222))
}
}
